fix(io): separate out io/ao contract interfaces
They are similar, but some differences. Separting them out will make it clear how they can be used
