+++ src/common/ar-io.ts
@@ -14,12 +14,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { ARNS_TESTNET_REGISTRY_TX, ioDevnetProcessId } from '../constants.js';
+import { ARNS_TESTNET_REGISTRY_TX } from '../constants.js';
 import {
   AR_IO_CONTRACT_FUNCTIONS,
-  AoEpochData,
-  AoGateway,
-  AoIOState,
   ArIOReadContract,
   ArIOState,
   ArIOWriteContract,
@@ -37,7 +34,6 @@ import {
   JoinNetworkParams,
   Observations,
   OptionalSigner,
-  ProcessConfiguration,
   RegistrationType,
   TransactionId,
   UpdateGatewaySettingsParams,
@@ -51,15 +47,9 @@ import {
 import {
   isContractConfiguration,
   isContractTxIdConfiguration,
-  isProcessConfiguration,
-  isProcessIdConfiguration,
 } from '../utils/smartweave.js';
 import { RemoteContract } from './contracts/remote-contract.js';
-import {
-  AOProcess,
-  InvalidContractConfigurationError,
-  WarpContract,
-} from './index.js';
+import { InvalidContractConfigurationError, WarpContract } from './index.js';
 
 export class ArIO {
   /**
@@ -206,9 +196,7 @@ export class ArIOReadable implements ArIOReadContract<ArIOState> {
    *  arIO.getState({ evaluationOptions: { evalTo: { sortKey: 'mySortKey' } } });
    * ```
    */
-  async getState<T = ArIOState | AoIOState>(
-    params: EvaluationParameters = {},
-  ): Promise<T> {
+  async getState<T = ArIOState>(params: EvaluationParameters = {}): Promise<T> {
     const state = await this.contract.getState(params);
     return state as T;
   }
@@ -1060,387 +1048,5 @@ export class ArIOWritable
       },
       options,
     );
-  }
-}
-
-// AO/IO Contract
-export class IO {
-  static init({ processId }: { processId: string }): IOReadable;
-  static init({ process }: { process: AOProcess<AoIOState> }): IOReadable;
-  static init(
-    config: WithSigner<
-      { process: AOProcess<AoIOState> } | { processId: string }
-    >,
-  ): IOWriteable;
-  static init(
-    config?: OptionalSigner<ProcessConfiguration<AoIOState>>,
-  ): IOReadable {
-    if (config && config.signer) {
-      const { signer, ...rest } = config;
-      return new IOWriteable({
-        ...rest,
-        signer,
-      });
-    }
-    return new IOReadable(config);
-  }
-}
-
-export class IOReadable
-  implements
-    Omit<
-      ArIOReadContract<AoIOState>,
-      'getAuction' | 'getAuctions' | 'getEpoch' | 'getPriceForInteraction'
-    >
-{
-  protected process: AOProcess<AoIOState>;
-
-  constructor(config?: ProcessConfiguration<AoIOState>) {
-    if (!config) {
-      this.process = new AOProcess<AoIOState>({
-        processId: ioDevnetProcessId,
-      });
-    } else if (isProcessConfiguration<AoIOState>(config)) {
-      this.process = config.process;
-    } else if (isProcessIdConfiguration(config)) {
-      this.process = new AOProcess<AoIOState>({
-        processId: config.processId,
-      });
-    } else {
-      throw new InvalidContractConfigurationError();
-    }
-  }
-
-  async getState(): Promise<AoIOState> {
-    return this.process.getState();
-  }
-
-  async getArNSRecord({
-    domain,
-  }: {
-    domain: string;
-  }): Promise<ArNSNameData | undefined> {
-    return this.process.read<ArNSNameData>({
-      tags: [
-        { name: 'Action', value: 'Record' },
-        { name: 'Name', value: domain },
-      ],
-    });
-  }
-
-  async getArNSRecords(): Promise<Record<string, ArNSNameData>> {
-    return this.process.read<Record<string, ArNSNameData>>({
-      tags: [{ name: 'Action', value: 'Records' }],
-    });
-  }
-
-  async getArNSReservedNames(): Promise<
-    Record<string, ArNSReservedNameData> | Record<string, never>
-  > {
-    return this.process.read<Record<string, ArNSReservedNameData>>({
-      tags: [{ name: 'Action', value: 'ReservedNames' }],
-    });
-  }
-
-  async getArNSReservedName({
-    domain,
-  }: {
-    domain: string;
-  }): Promise<ArNSReservedNameData | undefined> {
-    return this.process.read<ArNSReservedNameData>({
-      tags: [
-        { name: 'Action', value: 'ReservedName' },
-        { name: 'Name', value: domain },
-      ],
-    });
-  }
-
-  async getBalance({ address }: { address: WalletAddress }): Promise<number> {
-    return this.process.read<number>({
-      tags: [
-        { name: 'Action', value: 'Balance' },
-        { name: 'Address', value: address },
-      ],
-    });
-  }
-
-  async getBalances(): Promise<Record<WalletAddress, number>> {
-    return this.process.read<Record<string, number>>({
-      tags: [{ name: 'Action', value: 'Balances' }],
-    });
-  }
-
-  async getGateway({
-    address,
-  }: {
-    address: WalletAddress;
-  }): Promise<AoGateway | undefined> {
-    return this.process.read<AoGateway | undefined>({
-      tags: [
-        { name: 'Action', value: 'Gateway' },
-        { name: 'Address', value: address },
-      ],
-    });
-  }
-
-  async getGateways(): Promise<
-    Record<string, AoGateway> | Record<string, never>
-  > {
-    return this.process.read<Record<string, AoGateway>>({
-      tags: [{ name: 'Action', value: 'Gateways' }],
-    });
-  }
-
-  async getCurrentEpoch(): Promise<AoEpochData> {
-    return this.process.read<AoEpochData>({
-      tags: [
-        { name: 'Action', value: 'Epoch' },
-        { name: 'Timestamp', value: `${Date.now()}` },
-      ],
-    });
-  }
-
-  async getPrescribedObservers(): Promise<WeightedObserver[]> {
-    return this.process.read<WeightedObserver[]>({
-      tags: [{ name: 'Action', value: 'PrescribedObservers' }],
-    });
-  }
-
-  async getObservations(): Promise<Observations> {
-    return this.process.read<Observations>({
-      tags: [{ name: 'Action', value: 'Observations' }],
-    });
-  }
-
-  async getDistributions(): Promise<EpochDistributionData> {
-    return this.process.read<EpochDistributionData>({
-      tags: [{ name: 'Action', value: 'Distributions' }],
-    });
-  }
-}
-
-export class IOWriteable
-  extends IOReadable
-  implements
-    Omit<
-      ArIOReadContract<AoIOState>,
-      'getAuction' | 'getAuctions' | 'getEpoch' | 'getPriceForInteraction'
-    >
-{
-  protected declare process: AOProcess<AoIOState>;
-  private signer: ContractSigner;
-  constructor({
-    signer,
-    ...config
-  }: WithSigner<
-    | {
-        process?: AOProcess<AoIOState>;
-      }
-    | { processId?: string }
-  >) {
-    if (Object.keys(config).length === 0) {
-      super({
-        process: new AOProcess<AoIOState>({
-          processId: ioDevnetProcessId,
-        }),
-      });
-      this.signer = signer;
-    } else if (isProcessConfiguration<AoIOState>(config)) {
-      super({ process: config.process });
-      this.signer = signer;
-    } else if (isProcessIdConfiguration(config)) {
-      super({
-        process: new AOProcess<AoIOState>({
-          processId: config.processId,
-        }),
-      });
-      this.signer = signer;
-    } else {
-      throw new InvalidContractConfigurationError();
-    }
-  }
-
-  async transfer({
-    target,
-    qty,
-  }: {
-    target: string;
-    qty: number | mIOToken;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{
-      target: WalletAddress;
-      qty: number;
-      denomination?: DENOMINATIONS;
-    }>({
-      tags: [{ name: 'Action', value: 'Transfer' }],
-      data: {
-        target,
-        qty: qty.valueOf(),
-      },
-      signer: this.signer,
-    });
-  }
-
-  async joinNetwork({
-    qty,
-    allowDelegatedStaking,
-    delegateRewardShareRatio,
-    fqdn,
-    label,
-    minDelegatedStake,
-    note,
-    port,
-    properties,
-    protocol,
-    autoStake,
-    observerWallet,
-  }: JoinNetworkParams): Promise<WriteInteractionResult> {
-    return this.process.send<JoinNetworkParams>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'JoinNetwork' }],
-      data: {
-        qty: qty.valueOf(),
-        allowDelegatedStaking,
-        delegateRewardShareRatio,
-        fqdn,
-        label,
-        minDelegatedStake: minDelegatedStake.valueOf(),
-        note,
-        port,
-        properties,
-        protocol,
-        autoStake,
-        observerWallet,
-      },
-    });
-  }
-
-  async updateGatewaySettings({
-    allowDelegatedStaking,
-    delegateRewardShareRatio,
-    fqdn,
-    label,
-    minDelegatedStake,
-    note,
-    port,
-    properties,
-    protocol,
-    autoStake,
-    observerWallet,
-  }: UpdateGatewaySettingsParams): Promise<WriteInteractionResult> {
-    return this.process.send<UpdateGatewaySettingsParams>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'UpdateGatewaySettings' }],
-      data: {
-        allowDelegatedStaking,
-        delegateRewardShareRatio,
-        fqdn,
-        label,
-        minDelegatedStake: minDelegatedStake?.valueOf(),
-        note,
-        port,
-        properties,
-        protocol,
-        autoStake,
-        observerWallet,
-      },
-    });
-  }
-
-  async increaseDelegateStake(params: {
-    target: string;
-    qty: number | mIOToken;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ target: string; qty: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'IncreaseDelegateStake' }],
-      data: {
-        target: params.target,
-        qty: params.qty.valueOf(),
-      },
-    });
-  }
-
-  async decreaseDelegateStake(params: {
-    target: string;
-    qty: number | mIOToken;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ target: string; qty: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'DecreaseDelegateStake' }],
-      data: {
-        target: params.target,
-        qty: params.qty.valueOf(),
-      },
-    });
-  }
-
-  async increaseOperatorStake(params: {
-    qty: number | mIOToken;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ qty: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'IncreaseOperatorStake' }],
-      data: {
-        qty: params.qty.valueOf(),
-      },
-    });
-  }
-
-  async decreaseOperatorStake(params: {
-    qty: number | mIOToken;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ qty: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'DecreaseOperatorStake' }],
-      data: {
-        qty: params.qty.valueOf(),
-      },
-    });
-  }
-
-  async saveObservations(params: {
-    reportTxId: TransactionId;
-    failedGateways: WalletAddress[];
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{
-      observerReportTxId: TransactionId;
-      failedGateways: WalletAddress[];
-    }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'SaveObservations' }],
-      data: {
-        observerReportTxId: params.reportTxId,
-        failedGateways: params.failedGateways,
-      },
-    });
-  }
-
-  async extendLease(params: {
-    domain: string;
-    years: number;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ name: string; years: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'ExtendLease' }],
-      data: {
-        name: params.domain,
-        years: params.years,
-      },
-    });
-  }
-
-  async increaseUndernameLimit(params: {
-    domain: string;
-    qty: number;
-  }): Promise<WriteInteractionResult> {
-    return this.process.send<{ name: string; qty: number }>({
-      signer: this.signer,
-      tags: [{ name: 'Action', value: 'IncreaseUndernameLimit' }],
-      data: {
-        name: params.domain,
-        qty: params.qty,
-      },
-    });
   }
 }
